@@ -5,6 +5,7 @@ export default {
   ],
   components: [
     'pages/detail/index',
-    'components/picker/index'
+    'components/picker/index',
+    'components/article/index'
   ]
 }
