@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { View, Button } from '@tarojs/components'
 
 import './index.scss'
