@@ -1,11 +1,11 @@
 import * as assert from 'assert'
 import $ from 'jquery'
 import React from 'react'
-import ReactDOM from 'react-dom'
 import simulant from 'simulant'
 import * as sinon from 'sinon'
 
 import { Input } from '../h5/react'
+import { mount } from './test-tools'
 import { waitForChange } from './utils'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const h = React.createElement
@@ -85,7 +85,7 @@ describe('Input', () => {
       }
     }
 
-    ReactDOM.render(<App />, scratch)
+    await mount(<App />, scratch)
 
     /**
      * @type {HTMLElement}
@@ -153,7 +153,7 @@ describe('Input', () => {
       }
     }
 
-    ReactDOM.render(<App />, scratch)
+    await mount(<App />, scratch)
 
     /**
      * @type {HTMLElement}
