@@ -1,13 +1,29 @@
 import * as assert from 'assert'
 import React from 'react'
-import ReactDOM from 'react-dom'
 import * as sinon from 'sinon'
 
 import {
+  Ad,
+  AdCustom,
   Camera,
-  OpenData
+  Editor,
+  FunctionalPageNavigator,
+  KeyboardAccessory,
+  LivePlayer,
+  LivePusher,
+  Map,
+  MatchMedia,
+  NavigationBar,
+  OfficialAccount,
+  OpenData,
+  PageContainer,
+  PageMeta,
+  RootPortal,
+  ShareElement,
+  VoipRoom
 } from '../h5/react'
-import { waitForChange } from './utils'
+import { mount } from './test-tools'
+
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const h = React.createElement
 
@@ -43,15 +59,10 @@ describe('unimplemented', () => {
 
   async function testComponent (Comp) {
     const ref = React.createRef()
-    class App extends React.Component {
-      render () {
-        return <Comp ref={ref} />
-      }
-    }
-
-    ReactDOM.render(<App />, scratch)
-
-    await waitForChange(ref.current)
+    const app = (
+      <Comp ref={ref} />
+    )
+    await mount(app, scratch)
 
     assert(warning === buildWarning(ref))
   }
@@ -76,11 +87,75 @@ describe('unimplemented', () => {
     console.error.restore()
   })
 
-  it('OpenData', async () => {
-    await testComponent(OpenData)
+  it('Ad', async () => {
+    await testComponent(Ad)
+  })
+
+  it('AdCustom', async () => {
+    await testComponent(AdCustom)
   })
 
   it('Camera', async () => {
     await testComponent(Camera)
+  })
+
+  it('Editor', async () => {
+    await testComponent(Editor)
+  })
+
+  it('FunctionalPageNavigator', async () => {
+    await testComponent(FunctionalPageNavigator)
+  })
+
+  it('KeyboardAccessory', async () => {
+    await testComponent(KeyboardAccessory)
+  })
+
+  it('LivePlayer', async () => {
+    await testComponent(LivePlayer)
+  })
+
+  it('LivePusher', async () => {
+    await testComponent(LivePusher)
+  })
+
+  it('Map', async () => {
+    await testComponent(Map)
+  })
+
+  it('MatchMedia', async () => {
+    await testComponent(MatchMedia)
+  })
+
+  it('NavigationBar', async () => {
+    await testComponent(NavigationBar)
+  })
+
+  it('OfficialAccount', async () => {
+    await testComponent(OfficialAccount)
+  })
+
+  it('OpenData', async () => {
+    await testComponent(OpenData)
+  })
+
+  it('PageContainer', async () => {
+    await testComponent(PageContainer)
+  })
+
+  it('PageMeta', async () => {
+    await testComponent(PageMeta)
+  })
+
+  it('RootPortal', async () => {
+    await testComponent(RootPortal)
+  })
+
+  it('ShareElement', async () => {
+    await testComponent(ShareElement)
+  })
+
+  it('VoipRoom', async () => {
+    await testComponent(VoipRoom)
   })
 })
