@@ -1,12 +1,13 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Component, h, Host } from '@stencil/core'
+import { Component, ComponentInterface, h, Host } from '@stencil/core'
+
+import { notSupport } from '../../utils'
 
 @Component({
   tag: 'taro-functional-page-navigator-core'
 })
-export class FunctionalPageNavigator {
+export class FunctionalPageNavigator implements ComponentInterface {
   componentDidLoad () {
-    console.error('H5 暂不支持 FunctionalPageNavigator 组件！')
+    notSupport('FunctionalPageNavigator', this)
   }
 
   render () {
