@@ -1,12 +1,13 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Component, h, Host } from '@stencil/core'
+import { Component, ComponentInterface, h, Host } from '@stencil/core'
+
+import { notSupport } from '../../utils'
 
 @Component({
   tag: 'taro-open-data-core'
 })
-export class OpenData {
+export class OpenData implements ComponentInterface {
   componentDidLoad () {
-    console.error('H5 暂不支持 OpenData 组件！')
+    notSupport('OpenData', this)
   }
 
   render () {
