@@ -1,12 +1,13 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Component, h, Host } from '@stencil/core'
+import { Component, ComponentInterface, h, Host } from '@stencil/core'
+
+import { notSupport } from '../../utils'
 
 @Component({
   tag: 'taro-page-container-core'
 })
-export class PageContainer {
+export class PageContainer implements ComponentInterface {
   componentDidLoad () {
-    console.error('H5 暂不支持 PageContainer 组件！')
+    notSupport('PageContainer', this)
   }
 
   render () {
