@@ -3,48 +3,48 @@ import { StandardProps, CommonEventFunction } from './common'
 interface CameraProps extends StandardProps {
   /** 模式，有效值为normal, scanCode
    * @default "normal"
-   * @supported weapp, tt, qq, jd, rn, alipay
+   * @supported weapp, alipay, tt, qq, jd, rn
    */
   mode?: keyof CameraProps.Mode
 
   /** 分辨率，不支持动态修改
    * @default "medium"
-   * @supported weapp, tt, alipay
+   * @supported weapp, tt
    */
   resolution?: keyof CameraProps.Resolution
 
   /** 摄像头朝向
    * @default "back"
-   * @supported weapp, swan, tt, qq, jd, rn, alipay
+   * @supported weapp, alipay, swan, tt, qq, jd, rn
    */
   devicePosition?: keyof CameraProps.DevicePosition
 
   /** 闪光灯
    * @default "auto"
-   * @supported weapp, swan, tt, qq, jd, rn, alipay
+   * @supported weapp, alipay, swan, tt, qq, jd, rn
    */
   flash?: keyof CameraProps.Flash
 
-  /** 相机拍照，录制的分辨率。有效值为 360P、540P、720P、1080P、max
-   * @default "720P"
-   * @supported alipay
-   */
-  outputDimension?: '360P'|'540P'|'720P'|'1080P'|'max'
-
   /** 指定期望的相机帧数据尺寸
    * @default "medium"
-   * @supported weapp, tt, alipay
+   * @supported weapp, alipay, tt
    */
   frameSize?: keyof CameraProps.FrameSize
 
+  /** 相机拍照，录制的分辨率。
+   * @supported alipay
+   * @default "720P"
+   */
+  outputDimension?: '360P' | '540P' | '720P' | '1080P' | 'max'
+
   /** 摄像头在非正常终止时触发，
    * 如退出后台等情况
-   * @supported weapp, swan, tt, qq, jd, rn, alipay
+   * @supported weapp, alipay, swan, tt, qq, jd, rn
    */
   onStop?: CommonEventFunction
 
   /** 用户不允许使用摄像头时触发
-   * @supported weapp, swan, tt, qq, jd, rn, alipay
+   * @supported weapp, alipay, swan, tt, qq, jd, rn
    */
   onError?: CommonEventFunction
 
@@ -53,14 +53,14 @@ interface CameraProps extends StandardProps {
    */
   onInitDone?: CommonEventFunction<CameraProps.onInitDoneEventDetail>
 
-  /** 相机初始化完成时触发
+  /** 相机初始化成功时触发。
    * @supported alipay
    */
   onReady?: CommonEventFunction<CameraProps.onInitDoneEventDetail>
 
   /** 在成功识别到一维码时触发，
    * 仅在 mode="scanCode" 时生效
-   * @supported weapp, tt, qq, rn, alipay
+   * @supported weapp, alipay, tt, qq, rn
    */
   onScanCode?: CommonEventFunction<CameraProps.onScanCodeEventDetail>
 }
