@@ -164,17 +164,17 @@ interface ScrollViewProps extends StandardProps {
   /** 滑动开始事件 (同时开启 enhanced 属性后生效)
    * @supported weapp
    */
-  onDragStart?: CommonEventFunction
+  onDragStart?: CommonEventFunction<ScrollViewProps.onDragDetail>
 
   /** 滑动事件 (同时开启 enhanced 属性后生效)
    * @supported weapp
    */
-  onDragging?: CommonEventFunction
+  onDragging?: CommonEventFunction<ScrollViewProps.onDragDetail>
 
-  /** 滑动结束事件 (同时开启 enhanced 属性后生效) detail { scrollTop, scrollLeft, velocity }
+  /** 滑动结束事件 (同时开启 enhanced 属性后生效)
    * @supported weapp
    */
-  onDragend?: CommonEventFunction
+  onDragEnd?: CommonEventFunction<ScrollViewProps.onDragDetail>
 
   /** 当 scroll-with-animation设置为 true 时，可以设置 scroll-animation-duration 来控制动画的执行时间，单位 ms。
    * @supported alipay
@@ -221,6 +221,12 @@ interface ScrollViewProps extends StandardProps {
    * @supported qq
    */
   ariaLabel?: string
+
+  /** 开启 passive 特性，能优化一定的滚动性能
+   * @supported weapp
+   * @default false
+   */
+  enablePassive?: string
 }
 declare namespace ScrollViewProps {
   interface onScrollDetail {
@@ -237,6 +243,17 @@ declare namespace ScrollViewProps {
     scrollWidth: number
     deltaX: number
     deltaY: number
+  }
+
+  interface onDragDetail {
+    /** 横向滚动条位置 */
+    scrollLeft: number
+
+    /** 竖向滚动条位置 */
+    scrollTop: number
+
+    /** 滚动速度 */
+    velocity: number
   }
 }
 
