@@ -51,7 +51,7 @@ if (__TARO_FRAMEWORK__ === 'preact') {
     if (dom) { /** ElementNode */
       for (const propName in newVNodeProps) {
         const propValue = newVNodeProps[propName]
-        if (propValue === false && dom.props[propName] === undefined) {
+        if (propValue === false && dom.props?.[propName] === undefined) {
           // 值为 false 的属性在 Preact 的 diff 中被 removeAttribute 了，这里手动 setAttribute
           // fix https://github.com/NervJS/taro/issues/11197
           dom.setAttribute(propName, propValue)
