@@ -85,7 +85,8 @@ export class TaroRootElement extends TaroElement {
     const ctx = this.ctx!
 
     setTimeout(() => {
-      perf.start(SET_DATA)
+      const setDataMark = `${SET_DATA} 开始时间戳 ${Date.now()}`
+      perf.start(setDataMark)
       const data: Record<string, UpdatePayloadValue | ReturnType<HydratedData>> = Object.create(null)
       const resetPaths = new Set<string>(
         initRender
@@ -153,7 +154,7 @@ export class TaroRootElement extends TaroElement {
 
       const cb = () => {
         if (++executeTime === updateArrLen) {
-          perf.stop(SET_DATA)
+          perf.stop(setDataMark)
           this.flushUpdateCallback()
           initRender && perf.stop(PAGE_INIT)
         }
