@@ -1,7 +1,7 @@
 import { singleQuote } from '@tarojs/shared'
 
-const _true = 'true'
-const _false = 'false'
+const _true = '!0'
+const _false = '!1'
 const _empty = ''
 
 export const components = {
@@ -10,8 +10,8 @@ export const components = {
     size: '24'
   },
   Button: {
-    bindGetPhoneNumber: '',
-    'data-channel': ''
+    bindGetPhoneNumber: _empty,
+    'data-channel': _empty
   },
   Form: {
     'report-submit-timeout': '0'
@@ -49,20 +49,20 @@ export const components = {
     bindLeaveBackground: _empty,
   },
   Ad: {
-    fixed: '',
+    fixed: _empty,
     type: singleQuote('banner'),
     scale: '100'
   },
   Textarea: {
-    'disable-default-padding': 'false',
+    'disable-default-padding': _false,
     'confirm-type': singleQuote('return'),
-    'confirm-hold': 'false',
-    'show-confirm-bar': 'true',
-    'adjust-position': 'true',
-    'hold-keyboard': 'false'
+    'confirm-hold': _false,
+    'show-confirm-bar': _true,
+    'adjust-position': _true,
+    'hold-keyboard': _false
   },
   Canvas: {
-    type: ''
+    type: _empty
   },
   Map: {
     polygons: '[]',
@@ -70,26 +70,50 @@ export const components = {
     skew: '0',
     'max-scale': '19',
     'min-scale': '3',
-    'enable-3D': 'false',
-    'show-compass': 'false',
-    'show-scale': 'false',
-    'enable-overlooking': 'false',
-    'enable-zoom': 'true',
-    'enable-scroll': 'true',
-    'enable-rotate': 'false',
-    'enable-satellite': 'false',
-    'enable-traffic': 'false',
-    'enable-poi': 'true',
-    'enable-building': 'true',
-    bindLabelTap: '',
-    bindRegionChange: '',
-    bindAnchorPointTap: ''
+    'enable-3D': _false,
+    'show-compass': _false,
+    'show-scale': _false,
+    'enable-overlooking': _false,
+    'enable-zoom': _true,
+    'enable-scroll': _true,
+    'enable-rotate': _false,
+    'enable-satellite': _false,
+    'enable-traffic': _false,
+    'enable-poi': _true,
+    'enable-building': _true,
+    bindLabelTap: _empty,
+    bindRegionChange: _empty,
+    bindAnchorPointTap: _empty
   },
   // ======== 额外组件 ========
   RtcRoom: {
-    'user-id': '',
+    'user-id': _empty,
     mode: 'camera',
     'device-position': 'front',
-    bindError: ''
+    bindError: _empty
+  },
+  PayButton: {
+    mode: '1',
+    'goods-id': _empty,
+    'goods-type': _empty,
+    'order-status': '0',
+    'order-id': _empty,
+    'refund-id': _empty,
+    'refund-total-amount': _empty,
+    'biz-line': '1',
+    'marketing-ready': _false,
+    bindGetGoodsInfo: _empty,
+    bindPlaceOrder: _empty,
+    bindError: _empty,
+    bindApplyRefund: _empty,
+    bindRefund: _empty,
+    bindPay: _empty
+  },
+  ConsumeCard: {
+    'order-id': _empty,
+    bindConsume: _empty,
+    bindRefund: _empty,
+    bindApplyRefund: _empty,
+    bindError: _empty
   }
 }
