@@ -20,6 +20,7 @@ export interface IPrebundleParam {
   enableSourceMap?: boolean
   entryFileName?: string
 
+  isWatch?: boolean
   devServer?: webpackDevServer.Configuration
   publicPath?: string
   runtimePath?: string | string[]
@@ -43,6 +44,7 @@ export default class TaroPrebundle {
       enableSourceMap = false,
       entryFileName = 'app',
       entry = this.entry,
+      isWatch = false,
       publicPath = chain.output.get('publicPath') || '/',
       runtimePath,
       sourceRoot = 'src'
@@ -58,8 +60,9 @@ export default class TaroPrebundle {
       devServer,
       enableSourceMap,
       entryFileName,
-      entry: typeof entry === 'string' ? { [entryFileName]: entry } : entry,
+      entry,
       env,
+      isWatch,
       publicPath,
       runtimePath,
       sourceRoot
@@ -67,6 +70,7 @@ export default class TaroPrebundle {
   }
 
   get entry () {
+    // NOTE: 如果传入 entry 为字符串， webpack-chain 会识别为 EntryObject，导致报错
     return Object.entries(this.params.chain.entryPoints.entries()).reduce((entry, [key, value]) => {
       entry[key] = value.values()
       return entry
