@@ -95,7 +95,6 @@ export class MiniPrebundle extends BasePrebundle<IMiniPrebundleConfig> {
     const BUILD_LIB_START = performance.now()
 
     const exposes: Record<string, string> = {}
-    const mode = process.env.NODE_ENV === 'production' ? 'production' : 'development'
     const devtool = this.config.enableSourceMap && 'hidden-source-map'
     const mainBuildOutput = this.chain.output.entries()
     const taroRuntimeBundlePath: string = this.metadata.taroRuntimeBundlePath || exposes['./@tarojs/runtime']
@@ -124,7 +123,7 @@ export class MiniPrebundle extends BasePrebundle<IMiniPrebundleConfig> {
 
     this.metadata.mfHash = getMfHash({
       bundleHash: this.metadata.bundleHash,
-      mode,
+      mode: this.mode,
       devtool,
       output,
       taroRuntimeBundlePath
@@ -152,7 +151,7 @@ export class MiniPrebundle extends BasePrebundle<IMiniPrebundleConfig> {
         },
         devtool,
         entry: path.resolve(__dirname, './webpack/index.js'),
-        mode,
+        mode: this.mode,
         output,
         plugins: [
           new TaroModuleFederationPlugin(
