@@ -17,8 +17,12 @@ interface IH5PluginOptions {
   framework: FRAMEWORK_MAP
   frameworkExts: string[]
   useHtmlComponents: boolean
-  deviceRatio: any
-  designWidth: number
+  pxTransformConfig: {
+    baseFontSize: number
+    deviceRatio: any
+    designWidth: number
+    minRootSize: number
+  }
   prebundle?: boolean
   loaderMeta?: Record<string, string>
 }
@@ -40,8 +44,12 @@ export default class H5Plugin {
       framework: FRAMEWORK_MAP.NERV,
       frameworkExts: SCRIPT_EXT,
       useHtmlComponents: false,
-      deviceRatio: {},
-      designWidth: 750,
+      pxTransformConfig: {
+        baseFontSize: 20,
+        deviceRatio: {},
+        designWidth: 750,
+        minRootSize: 20
+      },
       prebundle: false
     })
   }
@@ -72,8 +80,8 @@ export default class H5Plugin {
     )
 
     compiler.hooks.compilation.tap(PLUGIN_NAME, compilation => {
-      compiler.webpack.NormalModule.getCompilationHooks(compilation).loader.tap(PLUGIN_NAME, (_loaderContext: LoaderContext<any>, module: NormalModule) => {
-        const { framework, entryFileName, appPath, sourceDir, designWidth, deviceRatio, loaderMeta, prebundle, routerConfig } = this.options
+      NormalModule.getCompilationHooks(compilation).loader.tap(PLUGIN_NAME, (_loaderContext: LoaderContext<any>, module: NormalModule) => {
+        const { framework, entryFileName, appPath, sourceDir, pxTransformConfig, loaderMeta, prebundle, routerConfig } = this.options
         const { dir, name } = path.parse(module.resource)
         const suffixRgx = /\.(boot|config)/
         if (!suffixRgx.test(name)) return
@@ -103,10 +111,7 @@ export default class H5Plugin {
               framework,
               loaderMeta,
               pages: this.inst.pagesConfigList,
-              pxTransformConfig: {
-                designWidth,
-                deviceRatio
-              },
+              pxTransformConfig,
               sourceDir,
               useHtmlComponents: this.options.useHtmlComponents
             },
